test(spec): add unit tests for OpenAI and Anthropic config specifications
